created user seed data
